restricted access for heads

+++ src/main/kotlin/com/meloda/kubsau/route/heads/HeadsRoutes.kt
@@ -1,13 +1,10 @@
 package com.meloda.kubsau.route.heads
 
-import com.meloda.kubsau.model.respondSuccess
 import com.meloda.kubsau.common.*
 import com.meloda.kubsau.database.faculties.FacultiesDao
 import com.meloda.kubsau.database.heads.HeadsDao
-import com.meloda.kubsau.model.ContentNotFoundException
-import com.meloda.kubsau.model.UnknownException
-import com.meloda.kubsau.model.Faculty
-import com.meloda.kubsau.model.Head
+import com.meloda.kubsau.model.*
+import com.meloda.kubsau.plugins.UserPrincipal
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
 import io.ktor.server.request.*
@@ -39,6 +36,8 @@ private fun Route.getHeads() {
     val facultiesDao by inject<FacultiesDao>()
 
     get {
+        val principal = call.principal<UserPrincipal>() ?: throw UnknownTokenException
+
         val parameters = call.request.queryParameters
 
         val headIds = parameters.getIntList(
@@ -52,9 +51,20 @@ private fun Route.getHeads() {
         val extended = parameters.getBoolean("extended", false)
 
         val heads = if (headIds.isEmpty()) {
-            headsDao.allHeads(offset, limit ?: MAX_ITEMS_SIZE)
+            headsDao.allHeads(
+                facultyId = principal.facultyId,
+                offset = offset,
+                limit = limit ?: MAX_ITEMS_SIZE
+            )
         } else {
             headsDao.allHeadsByIds(headIds)
+        }
+
+        if (principal.facultyId != null && headIds.isNotEmpty() &&
+            heads.map(Head::facultyId).distinct().singleOrNull() != principal.facultyId
+        ) {
+            val unavailableHeadIds = heads.filter { head -> head.facultyId != principal.facultyId }.map(Head::id)
+            throw AccessDeniedException("Unavailable headIds: ${unavailableHeadIds.joinToString()}")
         }
 
         if (!extended) {
@@ -87,10 +97,15 @@ private fun Route.getHeadById() {
     val facultiesDao by inject<FacultiesDao>()
 
     get("{id}") {
+        val principal = call.principal<UserPrincipal>() ?: throw UnknownTokenException
         val headId = call.parameters.getIntOrThrow("id")
         val extended = call.request.queryParameters.getBoolean("extended", false)
 
         val head = headsDao.singleHead(headId) ?: throw ContentNotFoundException
+
+        if (principal.facultyId != null && head.facultyId != principal.facultyId) {
+            throw AccessDeniedException("Unavailable headId: $headId")
+        }
 
         if (!extended) {
             respondSuccess { HeadResponse(head = head) }
@@ -111,12 +126,17 @@ private fun Route.addHead() {
     val headsDao by inject<HeadsDao>()
 
     post {
+        val principal = call.principal<UserPrincipal>() ?: throw UnknownTokenException
         val parameters = call.receiveParameters()
 
         val code = parameters.getStringOrThrow("code")
         val abbreviation = parameters.getStringOrThrow("abbreviation")
         val title = parameters.getStringOrThrow("title")
         val facultyId = parameters.getIntOrThrow("facultyId")
+
+        if (principal.facultyId != null && facultyId != principal.facultyId) {
+            throw AccessDeniedException("Unavailable facultyId: $facultyId")
+        }
 
         val created = headsDao.addNewHead(
             code = code,
@@ -137,6 +157,7 @@ private fun Route.editHead() {
     val headsDao by inject<HeadsDao>()
 
     patch("{id}") {
+        val principal = call.principal<UserPrincipal>() ?: throw UnknownTokenException
         val headId = call.parameters.getIntOrThrow("id")
         val currentHead = headsDao.singleHead(headId) ?: throw ContentNotFoundException
 
@@ -146,6 +167,10 @@ private fun Route.editHead() {
         val abbreviation = parameters.getString("abbreviation")
         val title = parameters.getString("title")
         val facultyId = parameters.getInt("facultyId")
+
+        if (principal.facultyId != null && facultyId != null && facultyId != principal.facultyId) {
+            throw AccessDeniedException("Unavailable facultyId: $facultyId")
+        }
 
         headsDao.updateHead(
             headId = headId,
@@ -167,8 +192,13 @@ private fun Route.deleteHead() {
     val headsDao by inject<HeadsDao>()
 
     delete("{id}") {
+        val principal = call.principal<UserPrincipal>() ?: throw UnknownTokenException
         val headId = call.parameters.getIntOrThrow("id")
-        headsDao.singleHead(headId) ?: throw ContentNotFoundException
+        val currentHead = headsDao.singleHead(headId) ?: throw ContentNotFoundException
+
+        if (principal.facultyId != null && currentHead.facultyId != principal.facultyId) {
+            throw AccessDeniedException("You can't delete this head with facultyId: ${currentHead.facultyId}")
+        }
 
         if (headsDao.deleteHead(headId)) {
             respondSuccess { 1 }
@@ -182,6 +212,7 @@ private fun Route.deleteHeads() {
     val headsDao by inject<HeadsDao>()
 
     delete {
+        val principal = call.principal<UserPrincipal>() ?: throw UnknownTokenException
         val headIds = call.request.queryParameters.getIntListOrThrow(
             key = "headIds",
             requiredNotEmpty = true
@@ -190,6 +221,13 @@ private fun Route.deleteHeads() {
         val currentHeads = headsDao.allHeadsByIds(headIds)
         if (currentHeads.isEmpty()) {
             throw ContentNotFoundException
+        }
+
+        if (principal.facultyId != null &&
+            currentHeads.map(Head::facultyId).distinct().singleOrNull() != principal.facultyId
+        ) {
+            val headsWithoutAccess = currentHeads.filter { head -> head.facultyId != principal.facultyId }.map(Head::id)
+            throw AccessDeniedException("You can't delete heads with ids: ${headsWithoutAccess.joinToString()}")
         }
 
         if (headsDao.deleteHeads(headIds)) {
